Add bitcoin onchain support scaffolding

+++ src/domains/bitcoin/entities/address.rs
@@ -8,7 +8,7 @@ pub struct BitcoinAddress {
     pub wallet_id: Uuid,
     pub address: String,
     pub used: bool,
-    pub derivation_index: Option<i64>,
+    pub derivation_index: Option<i32>,
     pub created_at: DateTime<Utc>,
     pub updated_at: Option<DateTime<Utc>>,
 }
